+++ internal/provider/data_source_list_all_system_roles_for_entra_group_test.go
@@ -15,7 +15,7 @@ func TestListAllSystemRolesForEntraGroupDataSource(t *testing.T) {
 	// A timestamp is added to the name to avoid failure due to previous test failures
 	time := time.Now().Unix()
 	roleName := fmt.Sprintf("TestListAllSystemRolesForEntraGroup Role %d", time)
-	groupName := fmt.Sprintf("[Test] group to be assigned TestListAllSystemRolesForEntraGroup %d", time)
+	groupName := fmt.Sprintf("[APPTEST] group to be assigned TestListAllSystemRolesForEntraGroup %d", time)
 	testUser := os.Getenv("ACC_TEST_SYSTEM_ROLE_OWNER")
 	itShopName := "General access shop shelf"
 
@@ -27,13 +27,13 @@ func TestListAllSystemRolesForEntraGroupDataSource(t *testing.T) {
 				Config: providerConfig + fmt.Sprintf(`
                 resource "tilgangsportalen_entra_group" "test_role_assignment_data_source" {
                     name = "%s"
-                    description = "terraform provider acceptance test"
+                    description = "APPTEST"
                     inheritance_level = "User"
                 }
 
                 data "tilgangsportalen_system_roles_assigned_to_entra_group" "roles_for_group" {
                     group_name = tilgangsportalen_entra_group.test_role_assignment_data_source.name
-					
+
                 }
                 `, groupName),
 				Check: resource.ComposeTestCheckFunc(
@@ -52,11 +52,11 @@ func TestListAllSystemRolesForEntraGroupDataSource(t *testing.T) {
                     approval_level    = "L2"
                     description       = "Terraform acceptance test role for assignment."
                     it_shop_name      = "%s"
-                } 
+                }
 
                 resource "tilgangsportalen_entra_group" "test_role_assignment_data_source" {
                     name = "%s"
-                    description = "terraform provider acceptance test"
+                    description = "APPTEST"
                     inheritance_level = "User"
                 }
 
@@ -67,7 +67,6 @@ func TestListAllSystemRolesForEntraGroupDataSource(t *testing.T) {
 
                 data "tilgangsportalen_system_roles_assigned_to_entra_group" "roles_for_group" {
                     group_name = tilgangsportalen_entra_group.test_role_assignment_data_source.name
-					
 					depends_on = [tilgangsportalen_entra_group_role_assignment.test_role_assignment_data_source]
                 }
                 `, roleName, testUser, itShopName, groupName),