+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.11.0
+
+ENHANCEMENTS:
+
+- Added return value `object_id` to resource `tilgangsportalen_entra_group`
+- Added return value `object_id` to data source `tilgangsportalen_entra_group`
+
 ## 0.10.1
 
 FIXES: